+++ packages/mask/src/extension/popups/pages/Wallet/Transfer/index.tsx
@@ -1,5 +1,6 @@
+import { FormattedBalance, TokenIcon, useMenu } from '@masknet/shared'
 import { NetworkPluginID } from '@masknet/shared-base'
-import { ActionButton, makeStyles } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import {
     useAccount,
     useBalance,
@@ -8,94 +9,35 @@ import {
     useNativeTokenAddress,
     useWallets,
 } from '@masknet/web3-hooks-base'
-import { explorerResolver, formatEthereumAddress, isNativeTokenAddress } from '@masknet/web3-shared-evm'
-import { Box, Link, List, ListItem, Typography, useTheme } from '@mui/material'
-import { memo, useMemo } from 'react'
+import { formatBalance, isSameAddress } from '@masknet/web3-shared-base'
+import { chainResolver, isNativeTokenAddress } from '@masknet/web3-shared-evm'
+import { MenuItem, Typography } from '@mui/material'
+import { first } from 'lodash-es'
+import { memo, useEffect, useMemo, useState } from 'react'
 import { useLocation, useParams } from 'react-router-dom'
+import { useContainer } from 'unstated-next'
 import { useI18N } from '../../../../../utils/index.js'
 import { useTitle } from '../../../hook/useTitle.js'
-import { ContactsContext } from './contactsContext.js'
-import { Icons } from '@masknet/icons'
-import { EmojiAvatar, FormattedAddress } from '@masknet/shared'
-import AddContactInputPanel from './AddContactInputPanel.js'
+import { WalletContext, useAsset } from '../hooks/index.js'
+import { Prior1559Transfer } from './Prior1559Transfer.js'
+import { Transfer1559 } from './Transfer1559.js'
 
-const useStyles = makeStyles()((theme) => ({
-    root: {
-        overflowX: 'hidden',
-        height: '100%',
-    },
-    page: {
-        position: 'relative',
-        height: '100%',
-        overflow: 'auto',
-    },
-    contactsPanel: {
-        display: 'flex',
-        flexDirection: 'column',
-        padding: '0 16px',
-    },
-    contactsList: {
-        padding: 0,
-    },
-    nickname: {
-        color: theme.palette.maskColor.main,
-        lineHeight: '18px',
-        fontWeight: 700,
-    },
-    identifier: {
-        fontSize: 14,
-        color: theme.palette.maskColor.second,
-        lineHeight: 1,
-        display: 'flex',
-        alignItems: 'center',
-    },
-    contactsListItem: {
+const useStyles = makeStyles()({
+    assetItem: {
         display: 'flex',
         justifyContent: 'space-between',
-        paddingLeft: '0px !important',
-        paddingRight: '0px !important',
+        minWidth: 278,
     },
-    contactsListItemInfo: {
+    assetSymbol: {
         display: 'flex',
         alignItems: 'center',
+        '& > p': {
+            marginLeft: 10,
+        },
     },
-    contactTitle: {
-        color: theme.palette.maskColor.main,
-        fontSize: 14,
-        fontWeight: 700,
-    },
-    icon: {
-        fontSize: 18,
-        height: 18,
-        width: 18,
-        color: theme.palette.maskColor.main,
-        cursor: 'pointer',
-        marginLeft: 4,
-    },
-    emojiAvatar: {
-        marginRight: 10,
-        fontSize: 14,
-    },
-    iconMore: {
-        cursor: 'pointer',
-    },
-    bottomAction: {
-        display: 'flex',
-        justifyContent: 'center',
-        background: theme.palette.maskColor.secondaryBottom,
-        boxShadow: '0px 0px 20px 0px rgba(0, 0, 0, 0.05)',
-        position: 'absolute',
-        backdropFilter: 'blur(8px)',
-        width: '100%',
-        bottom: 0,
-        zIndex: 100,
-    },
-    confirmButton: {
-        margin: '16px 0',
-    },
-}))
+})
 
-const TransferUI = memo(function TransferUI() {
+const Transfer = memo(function Transfer() {
     const location = useLocation()
     const { t } = useI18N()
     const { classes } = useStyles()
@@ -105,81 +47,67 @@ const TransferUI = memo(function TransferUI() {
     const isNativeToken = !address || isNativeTokenAddress(address)
     const account = useAccount(NetworkPluginID.PLUGIN_EVM)
     const wallets = useWallets(NetworkPluginID.PLUGIN_EVM)
-    const { receiver, setReceiver, receiverValidationMessage } = ContactsContext.useContainer()
-    const theme = useTheme()
-
+    const { assets } = useContainer(WalletContext)
     const { data: nativeTokenBalance = '0' } = useBalance(NetworkPluginID.PLUGIN_EVM)
     const { data: erc20Balance = '0' } = useFungibleTokenBalance(NetworkPluginID.PLUGIN_EVM, address)
     const balance = isNativeToken ? nativeTokenBalance : erc20Balance
+    const currentAsset = useAsset(chainId, address, account)
+    const [selectedAsset, setSelectedAsset] = useState(currentAsset ?? first(assets))
 
     const otherWallets = useMemo(
         () => wallets.map((wallet) => ({ name: wallet.name ?? '', address: wallet.address })),
         [wallets],
     )
 
+    const [assetsMenu, openAssetMenu] = useMenu(
+        ...assets.map((asset, index) => {
+            return (
+                <MenuItem key={index} className={classes.assetItem} onClick={() => setSelectedAsset(asset)}>
+                    <div className={classes.assetSymbol}>
+                        <TokenIcon
+                            chainId={asset.chainId}
+                            address={asset.address}
+                            name={asset.name}
+                            symbol={asset.symbol}
+                        />
+                        <Typography>{asset.symbol}</Typography>
+                    </div>
+                    <Typography>
+                        <FormattedBalance
+                            value={balance}
+                            decimals={asset.decimals}
+                            significant={4}
+                            formatter={formatBalance}
+                        />
+                    </Typography>
+                </MenuItem>
+            )
+        }),
+    )
+
     useTitle(t('popups_send'))
 
+    useEffect(() => {
+        const address = new URLSearchParams(location.search).get('selectedToken')
+        if (!address) return
+        const target = assets.find((x) => isSameAddress(x.address, address))
+        setSelectedAsset(target)
+    }, [assets, location])
+
     return (
-        <div className={classes.root}>
-            <Box className={classes.page}>
-                <AddContactInputPanel />
-                <Box className={classes.contactsPanel}>
-                    <Typography className={classes.contactTitle}>{t('wallet_transfer_my_wallets_title')}</Typography>
-                    <List className={classes.contactsList}>
-                        {wallets.map((wallet, index) => {
-                            return (
-                                <ListItem key={index} classes={{ root: classes.contactsListItem }}>
-                                    <div className={classes.contactsListItemInfo}>
-                                        <EmojiAvatar
-                                            address={wallet.address}
-                                            className={classes.emojiAvatar}
-                                            sx={{ width: 24, height: 24 }}
-                                        />
-                                        <div>
-                                            <Typography className={classes.nickname}>{wallet.name}</Typography>
-                                            <Typography className={classes.identifier}>
-                                                <FormattedAddress
-                                                    address={wallet.address}
-                                                    formatter={formatEthereumAddress}
-                                                    size={4}
-                                                />
-                                                <Link
-                                                    onClick={(event) => event.stopPropagation()}
-                                                    href={explorerResolver.addressLink(chainId, wallet.address ?? '')}
-                                                    target="_blank"
-                                                    rel="noopener noreferrer">
-                                                    <Icons.PopupLink className={classes.icon} />
-                                                </Link>
-                                            </Typography>
-                                        </div>
-                                    </div>
-                                    <Icons.More size={24} className={classes.iconMore} />
-                                </ListItem>
-                            )
-                        })}
-                    </List>
-                </Box>
-                <Box className={classes.bottomAction}>
-                    <ActionButton
-                        fullWidth
-                        onClick={() => {}}
-                        width={368}
-                        className={classes.confirmButton}
-                        disabled={!!receiverValidationMessage || !receiver}>
-                        {t('next')}
-                    </ActionButton>
-                </Box>
-            </Box>
-        </div>
+        <>
+            {chainResolver.isSupport(chainId, 'EIP1559') ? (
+                <Transfer1559 selectedAsset={selectedAsset} otherWallets={otherWallets} openAssetMenu={openAssetMenu} />
+            ) : (
+                <Prior1559Transfer
+                    selectedAsset={selectedAsset}
+                    otherWallets={otherWallets}
+                    openAssetMenu={openAssetMenu}
+                />
+            )}
+            {assetsMenu}
+        </>
     )
 })
-
-function Transfer() {
-    return (
-        <ContactsContext.Provider>
-            <TransferUI />
-        </ContactsContext.Provider>
-    )
-}
 
 export default Transfer
